backend/requirements.txt
fastapi==0.88.0
mangum==0.17.0
sqlalchemy==1.4.44
passlib==1.7.4
pymysql==1.0.2
bcrypt==4.0.1
boto3==1.26.27
python-multipart==0.0.5
python-jose==3.3.0
unique-names-generator==1.0.2
black==22.12.0
isort==5.11.1